# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/mcapi/health/data-enrich/ping": {
      "get": {
        "operationId": "HealthController_ping",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Basic health status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleHealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Basic health check endpoint",
        "tags": [
          "health"
        ]
      }
    },
    "/mcapi/health/data-enrich/status": {
      "get": {
        "operationId": "HealthController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Comprehensive health status with system metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Comprehensive health status endpoint",
        "tags": [
          "health"
        ]
      }
    },
    "/mcapi/data-enrich/buildings": {
      "get": {
        "description": "Retrieve buildings with comprehensive filtering options:\n    \n**Text Filtering:**\n- name: Partial match on building name (case-insensitive)\n- building_type: Exact match on building type (residential, commercial, industrial, mixed, other)\n- address: Partial match on building address (case-insensitive)\n- cadastral_code: Exact match on cadastral code\n- municipality_code: Exact match on municipality code\n\n**Geospatial Filtering:**\n- polygon: WKT polygon for spatial intersection queries (e.g., \"POLYGON((x1 y1, x2 y2, x3 y3, x1 y1))\")\n\n**Pagination:**\n- limit: Number of results per page (1-100, default: 20)\n- offset: Number of results to skip (default: 0)\n\nAll filters can be combined using AND logic. Only visible buildings are returned.",
        "operationId": "BuildingsController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip (default: 0)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results per page (1-100, default: 20)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "polygon",
            "required": false,
            "in": "query",
            "description": "WKT polygon for spatial filtering (e.g., \"POLYGON((x1 y1, x2 y2, x3 y3, x1 y1))\")",
            "schema": {}
          },
          {
            "name": "municipality_code",
            "required": false,
            "in": "query",
            "description": "Filter by municipality code (exact match)",
            "schema": {}
          },
          {
            "name": "cadastral_code",
            "required": false,
            "in": "query",
            "description": "Filter by cadastral code (exact match)",
            "schema": {}
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "description": "Filter by address (partial match)",
            "schema": {}
          },
          {
            "name": "building_type",
            "required": false,
            "in": "query",
            "description": "Filter by building type",
            "schema": {
              "enum": [
                "residential",
                "commercial",
                "industrial",
                "mixed",
                "other"
              ],
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by building name (partial match)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Buildings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBuildingsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters (validation errors)"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get buildings with filtering and pagination",
        "tags": [
          "buildings"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS API",
    "description": "Production-ready NestJS API with Fastify and Pino",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3232",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SimpleHealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "Basic health status",
            "type": "string"
          },
          "version": {
            "description": "Application version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "Health status of the application",
            "type": "string"
          },
          "timestamp": {
            "description": "ISO timestamp when health check was performed",
            "type": "string"
          },
          "server": {
            "description": "Server/application name",
            "type": "string"
          },
          "version": {
            "description": "Application version",
            "type": "string"
          },
          "uptime": {
            "description": "Server uptime in seconds",
            "type": "number"
          },
          "memory": {
            "description": "Memory usage statistics",
            "type": "object",
            "properties": {
              "rss": {
                "description": "Resident Set Size - total memory allocated",
                "type": "number"
              },
              "heapTotal": {
                "description": "Total heap memory allocated",
                "type": "number"
              },
              "heapUsed": {
                "description": "Heap memory currently used",
                "type": "number"
              },
              "external": {
                "description": "Memory used by C++ objects bound to JavaScript",
                "type": "number"
              },
              "arrayBuffers": {
                "description": "Memory allocated for ArrayBuffers and SharedArrayBuffers",
                "type": "number"
              }
            },
            "required": [
              "rss",
              "heapTotal",
              "heapUsed",
              "external",
              "arrayBuffers"
            ]
          },
          "environment": {
            "description": "Current environment (development, production, etc.)",
            "type": "string"
          }
        },
        "required": [
          "status",
          "timestamp",
          "server",
          "version",
          "uptime",
          "memory",
          "environment"
        ]
      },
      "PaginatedBuildingsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "cadastral_code": {
                  "type": "string"
                },
                "municipality_code": {
                  "type": "string"
                },
                "name": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "building_type": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "geom": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "const": "Point"
                        },
                        "coordinates": {
                          "type": "array",
                          "prefixItems": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "number"
                            }
                          ]
                        }
                      },
                      "required": [
                        "type",
                        "coordinates"
                      ]
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "basic_data": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                },
                "visible": {
                  "type": "boolean"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                },
                "updated_at": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "updated_by": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "cadastral_code",
                "municipality_code",
                "name",
                "building_type",
                "address",
                "geom",
                "basic_data",
                "visible",
                "created_at",
                "updated_at",
                "updated_by"
              ]
            },
            "x-nestjs_zod-has-null": true
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "offset": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "hasNext": {
                "type": "boolean"
              },
              "hasPrevious": {
                "type": "boolean"
              }
            },
            "required": [
              "total",
              "limit",
              "offset",
              "hasNext",
              "hasPrevious"
            ],
            "x-nestjs_zod-has-null": true
          }
        },
        "required": [
          "data",
          "meta"
        ]
      }
    }
  }
}